nova tabulka + vypis votes

+++ semestralka/app/Http/Controllers/TrackController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Track;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 
 class TrackController extends Controller
 {
@@ -18,19 +19,17 @@ class TrackController extends Controller
     {
         //$tracks = DB::select('select * from tracks', [1]);
         //$tracks = DB::select('select * from tracks where artist = :artist', ['artist' => 'Bran']);
-        $tracks = Track::all();
-        return view('track.allTracks', ['tracks' => $tracks]);
-
-        // if ($param == '1')
-        // {
-        //     $param = '1';
-        //     return view('track.allTracks', ['tracks' => $tracks, 'param' => $param]);
-        // }
-        // else {
-        //     $param = '0';
-        //     return view('track.allTracks', ['tracks' => $tracks, 'param' => $param]);
-        // }
-        
+        if (Auth::user() != null && Auth::user()->admin == 1)  
+        {
+            $tracks = Track::withCount('users')->get();
+            //$tracks = Track::all();
+            //dd($tracks->toArray());
+            return view('track.allTracks', ['tracks' => $tracks]);
+        }
+        else 
+        {
+            abort(403, 'Unauthorized access');        
+        }    
     }
 
     /**
@@ -40,10 +39,17 @@ class TrackController extends Controller
      */
     public function create()
     {
+        if (Auth::user() != null && Auth::user()->admin == 1)  
+        {
         return view('track.create', [
             'action' => route('track.store'),
             'method' => 'post'
         ]);
+        }
+        else 
+        {
+        abort(403, 'Unauthorized access');        
+        }
     }
 
     /**
@@ -54,6 +60,9 @@ class TrackController extends Controller
      */
     public function store(Request $request)
     {
+        if (Auth::user() != null && Auth::user()->admin == 1)  
+        {
+
         $request->validate([
             'artist' => 'required',
             'name' => 'required|min:6',
@@ -62,7 +71,13 @@ class TrackController extends Controller
 
         $track = Track::create($request->all());
         $track->save();
-        return redirect()->route('track.notify');    
+        return redirect()->route('track.notify');   
+        
+        }
+        else 
+        {
+        abort(403, 'Unauthorized access');        
+        } 
     }
 
     /**
@@ -81,7 +96,8 @@ class TrackController extends Controller
         //$tr = DB::update('update users set votes = 100 where name = ?', ['John']);
         // Track::delete(PK);
 
-        $tracks = DB::select('select * from tracks where genre = :genre order by votes desc', ['genre' => $genre]);
+        $tracks = Track::withCount('users')->orderBy('users_count', 'desc')->where('genre','=', $genre)->get();
+        //$tracks = DB::select('select * from tracks where genre = :genre order by votes desc', ['genre' => $genre]);
         return view('results.reslayout', ['tracks' => $tracks, 'genre' => $genre]);
     }
 
@@ -93,11 +109,18 @@ class TrackController extends Controller
      */
     public function edit(Track $track)
     {
+        if (Auth::user() != null && Auth::user()->admin == 1)  
+        {
         return view('track.edit', [
             'action' => route('track.update', $track->id),
             'method' => 'put',
             'model' => $track
         ]);
+        }
+        else 
+        {
+        abort(403, 'Unauthorized access');        
+        }
     }
 
     /**
@@ -109,6 +132,8 @@ class TrackController extends Controller
      */
     public function update(Request $request, Track $track)
     {
+        if (Auth::user() != null && Auth::user()->admin == 1)  
+        {
         $request->validate([
             'artist' => 'required',
             'name' => 'required|min:6',
@@ -116,7 +141,12 @@ class TrackController extends Controller
         ]);
 
         $track->update($request->all());
-        return redirect()->route('track.notify');    
+        return redirect()->route('track.notify');   
+        }
+        else 
+        {
+        abort(403, 'Unauthorized access');        
+        } 
     }
 
     /**
@@ -127,12 +157,26 @@ class TrackController extends Controller
      */
     public function destroy(Track $track)
     {
+        if (Auth::user() != null && Auth::user()->admin == 1)  
+        {
         $track->delete();
-        return redirect()->route('track.notify');    
+        return redirect()->route('track.notify');  
+        }
+        else 
+        {
+        abort(403, 'Unauthorized access');        
+        }  
     }
 
     public function notify(){
+        if (Auth::user() != null && Auth::user()->admin == 1)  
+        {
         $tracks = Track::all();
         return view('track.notify', ['tracks' => $tracks] );
+        }
+        else 
+        {
+        abort(403, 'Unauthorized access');        
+        }
     }
 }
